+++ DocuSignSDK.xcframework/ios-arm64_armv7/DocuSignSDK.framework/Headers/DSESSsn.h
@@ -16,7 +16,7 @@
 
 @interface DSESSsn : DSESObject
 
-/* A regular expressionn used to validate input for the tab. [optional]
+/* A regular expression used to validate input for the tab. [optional]
  */
 @property(nonatomic) NSString* validationPattern;
 /* The message displayed if the custom tab fails input validation (either custom of embedded). [optional]