+++ DocuSignSDK.xcframework/ios-arm64_armv7/DocuSignSDK.framework/Headers/DSESTabMetadata.h
@@ -105,7 +105,7 @@
 /* The message displayed if the custom tab fails input validation (either custom of embedded). [optional]
  */
 @property(nonatomic) NSString* validationMessage;
-/* A regular expressionn used to validate input for the tab. [optional]
+/* A regular expression used to validate input for the tab. [optional]
  */
 @property(nonatomic) NSString* validationPattern;
 /* The userId of the DocuSign user who last modified this object. [optional]