+++ DocuSignSDK.xcframework/ios-arm64_armv7/DocuSignSDK.framework/Headers/DSESUserSignatures.h
@@ -1,0 +1,15 @@
+#import <DocuSignSDK/DSESObject.h>
+#import <DocuSignSDK/DSESUserSignature.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@protocol DSESUserSignatures
+@end
+
+@interface DSESUserSignatures : DSESObject
+
+@property(nonatomic) NSArray<DSESUserSignature>* userSignatures;
+
+@end
+
+NS_ASSUME_NONNULL_END