+++ DocuSignSDK.xcframework/ios-arm64_armv7/DocuSignSDK.framework/Headers/DSMAccount.h
@@ -44,6 +44,7 @@ typedef enum {
 - (void)syncAdditionalDataWithCompletion:(DSMErrorBlock)completion;
 
 - (BOOL)canOfflineSend;
+- (BOOL)signerCanSignOnMobile;
 - (DSMSignAndReturnMethod)shouldSignAndReturnWithMethod;
 
 @end