+++ DocuSignSDK.xcframework/ios-arm64_armv7/DocuSignSDK.framework/Headers/DSMSigner.h
@@ -4,6 +4,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @class DSMTabs;
 @class DSMRecipientSignatureInformation;
+@class DSMConsumerDisclosure;
 
 /*!
  @class DSMSigner
@@ -68,6 +69,11 @@ NS_ASSUME_NONNULL_BEGIN
  * @brief TODO:Doc
  */
 @property(nonatomic, copy) NSString *emailRecipientPostSigningURL;
+
+/*!
+ * @brief Consumer Disclosure associated with the signer.
+ */
+@property(nonatomic, strong) DSMConsumerDisclosure *consumerDisclosure;
 
 @end