+++ DocuSignSDK.xcframework/ios-arm64_i386_x86_64-simulator/DocuSignSDK.framework/Headers/DSESConsumerDisclosure.h
@@ -5,5 +5,7 @@
 
 @property (nonatomic) NSString *accountEsignId;
 @property (nonatomic) NSString *esignAgreement;
+@property (nonatomic) NSString *mustAgreeToEsign;
+@property (nonatomic) NSString *companyName;
 
 @end