+++ DocuSignSDK.xcframework/ios-arm64_i386_x86_64-simulator/DocuSignSDK.framework/Headers/DSMText.h
@@ -16,7 +16,7 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) BOOL isPaymentAmount;
 /*! @brief The Formula string contains the TabLabel for the reference tabs used in the formula and calculation operators. Each TabLabel must be contained in brackets. \nMaximum Length: 2000 characters.\n\n*Example*: Three tabs (TabLabels: Line1, Line2, and Tax) need to be added together. The formula string would be: \n\n[Line1]+[Line2]+[Tax] [optional]. */
 @property (nonatomic, copy) NSString *formula;
-/*! @brief A regular expressionn used to validate input for the tab. [optional]. */
+/*! @brief A regular expression used to validate input for the tab. [optional]. */
 @property (nonatomic, copy) NSString *validationPattern;
 /*! @brief The message displayed if the custom tab fails input validation (either custom of embedded). [optional]. */
 @property (nonatomic, copy) NSString *validationMessage;