+++ DocuSignSDK.xcframework/ios-arm64_i386_x86_64-simulator/DocuSignSDK.framework/Headers/DSMUserSignature.h
@@ -1,7 +1,6 @@
 #import <Foundation/Foundation.h>
 #import <DocuSignSDK/DSESObject.h>
 
-
 @interface DSMUserSignature : DSESObject
 
 @property (nonatomic, copy) NSString *adoptedDateTime;